Moving Scenes to their own files -Created folder for scenes -Added files for each scene
